separated llm tracker by model
