+++ agentops/llms/openai.py
@@ -1,0 +1,358 @@
+import inspect
+import pprint
+from typing import Optional
+
+from agentops.time_travel import fetch_completion_override_from_time_travel_cache
+
+from agentops import LLMEvent, Session, ErrorEvent, logger
+from agentops.helpers import check_call_stack_for_agent_id, get_ISO_time
+
+
+def override_openai_v1_completion(tracker):
+    from openai.resources.chat import completions
+    from openai.types.chat import ChatCompletion, ChatCompletionChunk
+
+    # Store the original method
+    global original_create
+    original_create = completions.Completions.create
+
+    def patched_function(*args, **kwargs):
+        init_timestamp = get_ISO_time()
+        session = kwargs.get("session", None)
+        if "session" in kwargs.keys():
+            del kwargs["session"]
+
+        completion_override = fetch_completion_override_from_time_travel_cache(kwargs)
+        if completion_override:
+            result_model = None
+            pydantic_models = (ChatCompletion, ChatCompletionChunk)
+            for pydantic_model in pydantic_models:
+                try:
+                    result_model = pydantic_model.model_validate_json(
+                        completion_override
+                    )
+                    break
+                except Exception as e:
+                    pass
+
+            if result_model is None:
+                logger.error(
+                    f"Time Travel: Pydantic validation failed for {pydantic_models} \n"
+                    f"Time Travel: Completion override was:\n"
+                    f"{pprint.pformat(completion_override)}"
+                )
+                return None
+            return tracker._handle_response_v1_openai(
+                result_model, kwargs, init_timestamp, session=session
+            )
+
+        # prompt_override = fetch_prompt_override_from_time_travel_cache(kwargs)
+        # if prompt_override:
+        #     kwargs["messages"] = prompt_override["messages"]
+
+        # Call the original function with its original arguments
+        result = original_create(*args, **kwargs)
+        return tracker._handle_response_v1_openai(
+            result, kwargs, init_timestamp, session=session
+        )
+
+    # Override the original method with the patched one
+    completions.Completions.create = patched_function
+
+
+def override_openai_v1_async_completion(tracker):
+    from openai.resources.chat import completions
+    from openai.types.chat import ChatCompletion, ChatCompletionChunk
+
+    # Store the original method
+    global original_create_async
+    original_create_async = completions.AsyncCompletions.create
+
+    async def patched_function(*args, **kwargs):
+
+        init_timestamp = get_ISO_time()
+
+        session = kwargs.get("session", None)
+        if "session" in kwargs.keys():
+            del kwargs["session"]
+
+        completion_override = fetch_completion_override_from_time_travel_cache(kwargs)
+        if completion_override:
+            result_model = None
+            pydantic_models = (ChatCompletion, ChatCompletionChunk)
+            for pydantic_model in pydantic_models:
+                try:
+                    result_model = pydantic_model.model_validate_json(
+                        completion_override
+                    )
+                    break
+                except Exception as e:
+                    pass
+
+            if result_model is None:
+                logger.error(
+                    f"Time Travel: Pydantic validation failed for {pydantic_models} \n"
+                    f"Time Travel: Completion override was:\n"
+                    f"{pprint.pformat(completion_override)}"
+                )
+                return None
+            return tracker._handle_response_v1_openai(
+                result_model, kwargs, init_timestamp, session=session
+            )
+
+        # prompt_override = fetch_prompt_override_from_time_travel_cache(kwargs)
+        # if prompt_override:
+        #     kwargs["messages"] = prompt_override["messages"]
+
+        # Call the original function with its original arguments
+        result = await original_create_async(*args, **kwargs)
+        return tracker._handle_response_v1_openai(
+            result, kwargs, init_timestamp, session=session
+        )
+
+    # Override the original method with the patched one
+    completions.AsyncCompletions.create = patched_function
+
+
+def handle_response_v0_openai(
+    tracker, response, kwargs, init_timestamp, session: Optional[Session] = None
+):
+    """Handle responses for OpenAI versions <v1.0.0"""
+
+    tracker.llm_event = LLMEvent(init_timestamp=init_timestamp, params=kwargs)
+    if session is not None:
+        tracker.llm_event.session_id = session.session_id
+
+    def handle_stream_chunk(chunk):
+        # NOTE: prompt/completion usage not returned in response when streaming
+        # We take the first ChatCompletionChunk and accumulate the deltas from all subsequent chunks to build one full chat completion
+        if tracker.llm_event.returns == None:
+            tracker.llm_event.returns = chunk
+
+        try:
+            accumulated_delta = tracker.llm_event.returns["choices"][0]["delta"]
+            tracker.llm_event.agent_id = check_call_stack_for_agent_id()
+            tracker.llm_event.model = chunk["model"]
+            tracker.llm_event.prompt = kwargs["messages"]
+            choice = chunk["choices"][
+                0
+            ]  # NOTE: We assume for completion only choices[0] is relevant
+
+            if choice["delta"].get("content"):
+                accumulated_delta["content"] += choice["delta"].content
+
+            if choice["delta"].get("role"):
+                accumulated_delta["role"] = choice["delta"].get("role")
+
+            if choice["finish_reason"]:
+                # Streaming is done. Record LLMEvent
+                tracker.llm_event.returns.choices[0]["finish_reason"] = choice[
+                    "finish_reason"
+                ]
+                tracker.llm_event.completion = {
+                    "role": accumulated_delta["role"],
+                    "content": accumulated_delta["content"],
+                }
+                tracker.llm_event.end_timestamp = get_ISO_time()
+
+                tracker._safe_record(session, tracker.llm_event)
+        except Exception as e:
+            tracker._safe_record(
+                session, ErrorEvent(trigger_event=tracker.llm_event, exception=e)
+            )
+
+            kwargs_str = pprint.pformat(kwargs)
+            chunk = pprint.pformat(chunk)
+            logger.warning(
+                f"Unable to parse a chunk for LLM call. Skipping upload to AgentOps\n"
+                f"chunk:\n {chunk}\n"
+                f"kwargs:\n {kwargs_str}\n"
+            )
+
+    # if the response is a generator, decorate the generator
+    if inspect.isasyncgen(response):
+
+        async def async_generator():
+            async for chunk in response:
+                handle_stream_chunk(chunk)
+
+                yield chunk
+
+        return async_generator()
+
+    elif inspect.isgenerator(response):
+
+        def generator():
+            for chunk in response:
+                handle_stream_chunk(chunk)
+
+                yield chunk
+
+        return generator()
+
+    # v0.0.0 responses are dicts
+    try:
+        tracker.llm_event.returns = response
+        tracker.llm_event.agent_id = check_call_stack_for_agent_id()
+        tracker.llm_event.prompt = kwargs["messages"]
+        tracker.llm_event.prompt_tokens = response["usage"]["prompt_tokens"]
+        tracker.llm_event.completion = {
+            "role": "assistant",
+            "content": response["choices"][0]["message"]["content"],
+        }
+        tracker.llm_event.completion_tokens = response["usage"]["completion_tokens"]
+        tracker.llm_event.model = response["model"]
+        tracker.llm_event.end_timestamp = get_ISO_time()
+
+        tracker._safe_record(session, tracker.llm_event)
+    except Exception as e:
+        tracker._safe_record(
+            session, ErrorEvent(trigger_event=tracker.llm_event, exception=e)
+        )
+
+        kwargs_str = pprint.pformat(kwargs)
+        response = pprint.pformat(response)
+        logger.warning(
+            f"Unable to parse response for LLM call. Skipping upload to AgentOps\n"
+            f"response:\n {response}\n"
+            f"kwargs:\n {kwargs_str}\n"
+        )
+
+    return response
+
+
+def handle_response_v1_openai(
+    tracker, response, kwargs, init_timestamp, session: Optional[Session] = None
+):
+    """Handle responses for OpenAI versions >v1.0.0"""
+    from openai import AsyncStream, Stream
+    from openai.resources import AsyncCompletions
+    from openai.types.chat import ChatCompletionChunk
+
+    tracker.llm_event = LLMEvent(init_timestamp=init_timestamp, params=kwargs)
+    if session is not None:
+        tracker.llm_event.session_id = session.session_id
+
+    def handle_stream_chunk(chunk: ChatCompletionChunk):
+        # NOTE: prompt/completion usage not returned in response when streaming
+        # We take the first ChatCompletionChunk and accumulate the deltas from all subsequent chunks to build one full chat completion
+        if tracker.llm_event.returns == None:
+            tracker.llm_event.returns = chunk
+
+        try:
+            accumulated_delta = tracker.llm_event.returns.choices[0].delta
+            tracker.llm_event.agent_id = check_call_stack_for_agent_id()
+            tracker.llm_event.model = chunk.model
+            tracker.llm_event.prompt = kwargs["messages"]
+
+            # NOTE: We assume for completion only choices[0] is relevant
+            choice = chunk.choices[0]
+
+            if choice.delta.content:
+                accumulated_delta.content += choice.delta.content
+
+            if choice.delta.role:
+                accumulated_delta.role = choice.delta.role
+
+            if choice.delta.tool_calls:
+                accumulated_delta.tool_calls = choice.delta.tool_calls
+
+            if choice.delta.function_call:
+                accumulated_delta.function_call = choice.delta.function_call
+
+            if choice.finish_reason:
+                # Streaming is done. Record LLMEvent
+                tracker.llm_event.returns.choices[0].finish_reason = (
+                    choice.finish_reason
+                )
+                tracker.llm_event.completion = {
+                    "role": accumulated_delta.role,
+                    "content": accumulated_delta.content,
+                    "function_call": accumulated_delta.function_call,
+                    "tool_calls": accumulated_delta.tool_calls,
+                }
+                tracker.llm_event.end_timestamp = get_ISO_time()
+
+                tracker._safe_record(session, tracker.llm_event)
+        except Exception as e:
+            tracker._safe_record(
+                session, ErrorEvent(trigger_event=tracker.llm_event, exception=e)
+            )
+
+            kwargs_str = pprint.pformat(kwargs)
+            chunk = pprint.pformat(chunk)
+            logger.warning(
+                f"Unable to parse a chunk for LLM call. Skipping upload to AgentOps\n"
+                f"chunk:\n {chunk}\n"
+                f"kwargs:\n {kwargs_str}\n"
+            )
+
+    # if the response is a generator, decorate the generator
+    if isinstance(response, Stream):
+
+        def generator():
+            for chunk in response:
+                handle_stream_chunk(chunk)
+                yield chunk
+
+        return generator()
+
+    # For asynchronous AsyncStream
+    elif isinstance(response, AsyncStream):
+
+        async def async_generator():
+            async for chunk in response:
+                handle_stream_chunk(chunk)
+                yield chunk
+
+        return async_generator()
+
+    # For async AsyncCompletion
+    elif isinstance(response, AsyncCompletions):
+
+        async def async_generator():
+            async for chunk in response:
+                handle_stream_chunk(chunk)
+                yield chunk
+
+        return async_generator()
+
+    # v1.0.0+ responses are objects
+    try:
+        tracker.llm_event.returns = response
+        tracker.llm_event.agent_id = check_call_stack_for_agent_id()
+        tracker.llm_event.prompt = kwargs["messages"]
+        tracker.llm_event.prompt_tokens = response.usage.prompt_tokens
+        tracker.llm_event.completion = response.choices[0].message.model_dump()
+        tracker.llm_event.completion_tokens = response.usage.completion_tokens
+        tracker.llm_event.model = response.model
+
+        tracker._safe_record(session, tracker.llm_event)
+    except Exception as e:
+        tracker._safe_record(
+            session, ErrorEvent(trigger_event=tracker.llm_event, exception=e)
+        )
+
+        kwargs_str = pprint.pformat(kwargs)
+        response = pprint.pformat(response)
+        logger.warning(
+            f"Unable to parse response for LLM call. Skipping upload to AgentOps\n"
+            f"response:\n {response}\n"
+            f"kwargs:\n {kwargs_str}\n"
+        )
+
+    return response
+
+
+def undo_override_openai_v1_completion():
+    global original_create
+    from openai.resources.chat import completions
+
+    completions.Completions.create = original_create
+
+
+def undo_override_openai_v1_async_completion():
+    global original_create_async
+    from openai.resources.chat import completions
+
+    completions.AsyncCompletions.create = original_create_async
